server/python/requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.2
aniso8601==1.3.0
click==6.7
itsdangerous==0.24
mysqlclient==1.3.12
paho-mqtt==1.3.1
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
